rename: change interfaces file to types file
